backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1


httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1


playwright==1.44.0


playwright-stealth==1.0.6